Restructure things
